+++ Staging/AudioDxe/AudioDxe.c
@@ -44,6 +44,9 @@ BOOLEAN
 EFI_DEVICE_PATH_PROTOCOL *
   gForcedControllerDevicePath = NULL;
 
+UINTN
+  gCodecSetupDelay = 0;
+
 /**
   HdaController Driver Binding.
 **/
@@ -95,6 +98,8 @@ AudioDxeInit (
     return Status;
   }
 
+  DevicePathName = NULL;
+
   Status = OcParseLoadOptions (LoadedImage, &ParsedLoadOptions);
   if (!EFI_ERROR (Status)) {
     gRestoreNoSnoop = OcHasParsedVar (ParsedLoadOptions, L"--restore-nosnoop", OcStringFormatUnicode);
@@ -104,29 +109,32 @@ AudioDxeInit (
       gGpioSetupStageMask = GPIO_SETUP_STAGE_ALL;
     }
 
-    DEBUG ((DEBUG_INFO, "HDA: GPIO setup stages 0x%X, restore NSNPEN %d\n", gGpioSetupStageMask, gRestoreNoSnoop));
-
     if (gGpioSetupStageMask != GPIO_SETUP_STAGE_NONE) {
       OcParsedVarsGetInt (ParsedLoadOptions, L"--gpio-pins", &gGpioPinMask, OcStringFormatUnicode);
-      DEBUG ((
-        DEBUG_INFO,
-        "HDA: GPIO pin mask 0x%X%a\n",
-        gGpioPinMask,
-        gGpioPinMask == 0 ? " (auto)" : ""
-        ));
     }
 
-    DevicePathName = NULL;
     OcParsedVarsGetUnicodeStr (ParsedLoadOptions, L"--force-device", &DevicePathName);
     if (DevicePathName != NULL) {
-      DEBUG ((DEBUG_INFO, "HDA: Force binding to %s\n", DevicePathName));
       gForcedControllerDevicePath = ConvertTextToDevicePath (DevicePathName);
     }
+
+    OcParsedVarsGetInt (ParsedLoadOptions, L"--codec-setup-delay", &gCodecSetupDelay, OcStringFormatUnicode);
 
     OcFlexArrayFree (&ParsedLoadOptions);
   } else if (Status != EFI_NOT_FOUND) {
     return Status;
   }
+
+  DEBUG ((
+    DEBUG_INFO,
+    "HDA: GPIO setup stages 0x%X GPIO pin mask 0x%X%a Restore NSNPEN %d Force device %s Setup delay %u\n",
+    gGpioSetupStageMask,
+    gGpioPinMask,
+    gGpioPinMask == 0 ? " (auto)" : "",
+    gRestoreNoSnoop,
+    DevicePathName,
+    gCodecSetupDelay
+    ));
 
   //
   // Register HdaController Driver Binding.